Move internal/workers to /workers
