+++ workers/gbfs_fetch_worker.go
