+++ workers/gbfs_fetch_worker_test.go
