+++ workers/static_fetch_worker.go
